% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example}
\alias{example}
\title{Example Function}
\usage{
example()
}
\description{
Example Function
}
\examples{
ncores <- Sys.getenv("OMP_NUM_THREADS", 2)
Sys.setenv("OMP_NUM_THREADS"=2)
example()
Sys.setenv("OMP_NUM_THREADS"=ncores)

}
