% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{openmp_example}
\alias{openmp_example}
\title{Example Function to Illustrate OpenMP Use}
\usage{
openmp_example()
}
\description{
This example is taken directly from the Armadillo documentation.
When built with OpenMP support, and running on a system with
multiple cores, R will likely complain about 'CPU time > 2 elapsed
time'. However, when environment variable 'SHOWME' is set to 'yes'
we throttle and no warning appears.
}
\examples{
if (Sys.getenv("SHOWME", "") == "yes")
   set_number_of_omp_threads(1)
openmp_example()

}
