\name{RcppArmadilloOpenMPEx-package}
\alias{RcppArmadilloOpenMPEx-package}
\alias{RcppArmadilloOpenMPEx}
\docType{package}
\title{\packageTitle{RcppArmadilloOpenMPEx}}
\description{\packageDescription{RcppArmadilloOpenMPEx}}
\section{Package Content}{\packageIndices{RcppArmadilloOpenMPEx}}
\author{\packageAuthor{RcppArmadilloOpenMPEx}}
\section{Maintainer}{\packageMaintainer{RcppArmadilloOpenMPEx}}
\keyword{package}
