% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_number_of_omp_threads}
\alias{set_number_of_omp_threads}
\title{Set Maximum Number of Threads}
\usage{
set_number_of_omp_threads(n)
}
\arguments{
\item{n}{Number of threads to be set}
}
\description{
Set Maximum Number of Threads
}
